requirements.txt
python-dotenv
python-jose
bcrypt==4.0.1
passlib[bcrypt]
pydantic-settings
pydantic[email]
sqlalchemy[asyncio]
alembic
asyncpg
fastapi